+++ packages/dynamics/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-dynamics
 
+## 0.5.6
+
+### Patch Changes
+
+- 8d866e4: Update tough-cookie dependency
+
 ## 0.5.5
 
 ### Patch Changes